Cleanup, add binding for rating on details view
