+++ SimpleRss/SimpleRss/Extensions/UIImageView+Extension.swift
@@ -1,0 +1,16 @@
+//
+//  UIImageView+Extension.swift
+//  SimpleRss
+//
+//  Created by Voldem on 12/26/19.
+//  Copyright © 2019 Vladimir Koptev. All rights reserved.
+//
+
+import Foundation
+
+extension UIImageView {
+    
+    static var urlToImagePlaceholder: URL {
+        return URL(string: "https://aliceasmartialarts.com/wp-content/uploads/2017/04/default-image.jpg")!
+    }
+}
