+++ SimpleRss/SimpleRss/ViewControllers/FeedDetail.storyboard
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="15400" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="84V-Dm-Lez">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="15505" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="84V-Dm-Lez">
     <device id="retina4_7" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15404"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15510"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
@@ -59,7 +59,7 @@
                                                         <color key="textColor" white="0.66666666669999997" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                                         <nil key="highlightedColor"/>
                                                     </label>
-                                                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="n96-HS-qZe" customClass="RatingView" customModule="SimpleRss" customModuleProvider="target">
+                                                    <view userInteractionEnabled="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="n96-HS-qZe" customClass="RatingView" customModule="SimpleRss" customModuleProvider="target">
                                                         <rect key="frame" x="8" y="31.5" width="359" height="44"/>
                                                         <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                                         <constraints>
@@ -113,6 +113,7 @@
                         <outlet property="collectionView" destination="9ST-jP-DLX" id="r67-Me-2lD"/>
                         <outlet property="descriptionTextView" destination="skP-GB-cgB" id="IE6-zn-Fe8"/>
                         <outlet property="pubDateLabel" destination="bfN-UC-qlS" id="xQV-0d-rLL"/>
+                        <outlet property="ratingView" destination="n96-HS-qZe" id="l5L-Sq-mde"/>
                         <outlet property="titleLabel" destination="lmL-L1-EPQ" id="T9p-Le-wYW"/>
                     </connections>
                 </viewController>
